+++ EventSubscriber/EnumEventSubscriber.php
@@ -5,20 +5,21 @@ declare(strict_types=1);
 /*
  * This file is part of the DbalEnumType package.
  *
- * (c) Wakeapp <https://wakeapp.ru>
+ * (c) MarfaTech <https://marfa-tech.com>
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
 
-namespace Wakeapp\Component\DbalEnumType\EventSubscriber;
+namespace MarfaTech\Component\DbalEnumType\EventSubscriber;
 
 use Doctrine\Common\EventSubscriber;
 use Doctrine\DBAL\Event\SchemaColumnDefinitionEventArgs;
 use Doctrine\DBAL\Events;
-use Wakeapp\Component\DbalEnumType\Exception\EnumException;
-use Wakeapp\Component\DbalEnumType\Schema\EnumAwareMySqlSchemaManager;
-use Wakeapp\Component\DbalEnumType\Type\AbstractEnumType;
+use MarfaTech\Component\DbalEnumType\Exception\EnumException;
+use MarfaTech\Component\DbalEnumType\Schema\EnumAwareMySqlSchemaManager;
+use MarfaTech\Component\DbalEnumType\Type\AbstractEnumType;
+
 use function mb_strtolower;
 
 class EnumEventSubscriber implements EventSubscriber