+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 ## [Unreleased]
+## Changed
+- Updated `php` with pattern version `~7.4||~8.0`.
+- Updated composer name `wakeapp/dbal-enum-type` to `marfatech/dbal-enum-type`.
+- [BC BREAK] Refactoring namespace to `MarfaTech`.
 
 ## [1.0.1] - 2021-03-02
 ### Added